+++ packages/core/src/lib/light_push/index.ts
@@ -47,7 +47,7 @@ class LightPush extends BaseProtocol implements ILightPush {
   private readonly pubsubTopics: PubsubTopic[];
 
   constructor(libp2p: Libp2p, options?: ProtocolCreateOptions) {
-    super(LightPushCodec, libp2p.components);
+    super(LightPushCodec, libp2p.components, log);
     this.pubsubTopics = this.initializePubsubTopic(options);
   }