+++ packages/core/src/lib/store/index.ts
@@ -78,7 +78,7 @@ class Store extends BaseProtocol implements IStore {
   private readonly NUM_PEERS_PROTOCOL = 1;
 
   constructor(libp2p: Libp2p, options?: ProtocolCreateOptions) {
-    super(StoreCodec, libp2p.components);
+    super(StoreCodec, libp2p.components, log);
     this.pubsubTopics = this.initializePubsubTopic(options);
   }