+++ packages/tests/tests/filter/utils.ts
@@ -12,7 +12,7 @@ import { Logger } from "@waku/utils";
 import { utf8ToBytes } from "@waku/utils/bytes";
 import { Context } from "mocha";
 
-import { makeLogFileName, NimGoNode, NOISE_KEY_1 } from "../../src/index.js";
+import { makeLogFileName, NOISE_KEY_1, ServiceNode } from "../../src/index.js";
 
 // Constants for test configuration.
 export const log = new Logger("test:filter");
@@ -48,8 +48,8 @@ export async function runNodes(
   //TODO: change this to use `ShardInfo` instead of `string[]`
   pubsubTopics: string[],
   shardInfo?: ShardingParams
-): Promise<[NimGoNode, LightNode]> {
-  const nwaku = new NimGoNode(makeLogFileName(context));
+): Promise<[ServiceNode, LightNode]> {
+  const nwaku = new ServiceNode(makeLogFileName(context));
 
   await nwaku.start(
     {