+++ packages/tests/tests/peer_exchange.node.spec.ts
@@ -1,4 +1,6 @@
+import type { PeerId } from "@libp2p/interface/peer-id";
 import tests from "@libp2p/interface-compliance-tests/peer-discovery";
+import type { Multiaddr } from "@multiformats/multiaddr";
 import type { LightNode, PeerInfo } from "@waku/interfaces";
 import {
   PeerExchangeCodec,
@@ -8,20 +10,23 @@ import {
 import { createLightNode, Libp2pComponents } from "@waku/sdk";
 import { expect } from "chai";
 
-import { delay } from "../src/delay.js";
-import { tearDownNodes } from "../src/index.js";
-import { makeLogFileName } from "../src/log_file.js";
-import { NimGoNode } from "../src/node/node.js";
+import {
+  delay,
+  makeLogFileName,
+  ServiceNode,
+  tearDownNodes,
+  waitForRemotePeerWithCodec
+} from "../src/index.js";
 
 describe("Peer Exchange", () => {
   describe("Locally Run Nodes", () => {
     let waku: LightNode;
-    let nwaku1: NimGoNode;
-    let nwaku2: NimGoNode;
+    let nwaku1: ServiceNode;
+    let nwaku2: ServiceNode;
 
     beforeEach(function () {
-      nwaku1 = new NimGoNode(makeLogFileName(this) + "1");
-      nwaku2 = new NimGoNode(makeLogFileName(this) + "2");
+      nwaku1 = new ServiceNode(makeLogFileName(this) + "1");
+      nwaku2 = new ServiceNode(makeLogFileName(this) + "2");
     });
 
     afterEach(async function () {
@@ -48,11 +53,13 @@ describe("Peer Exchange", () => {
       });
 
       const nwaku1PeerId = await nwaku1.getPeerId();
+      const nwaku2PeerId = await nwaku2.getPeerId();
       const nwaku2Ma = await nwaku2.getMultiaddrWithId();
 
       waku = await createLightNode();
       await waku.start();
       await waku.libp2p.dialProtocol(nwaku2Ma, PeerExchangeCodec);
+      await waitForRemotePeerWithCodec(waku, PeerExchangeCodec, nwaku2PeerId);
 
       const components = waku.libp2p.components as unknown as Libp2pComponents;
       const peerExchange = new WakuPeerExchange(components);
@@ -62,6 +69,7 @@ describe("Peer Exchange", () => {
       let peerInfos: PeerInfo[] = [];
       while (peerInfos.length <= 0) {
         peerInfos = (await peerExchange.query({
+          peerId: nwaku2PeerId,
           numPeers: numPeersToRequest
         })) as PeerInfo[];
         await delay(3000);
@@ -70,16 +78,35 @@ describe("Peer Exchange", () => {
       expect(peerInfos.length).to.be.greaterThan(0);
       expect(peerInfos.length).to.be.lessThanOrEqual(numPeersToRequest);
       expect(peerInfos[0].ENR).to.not.be.null;
+      expect(peerInfos[0].ENR?.peerInfo?.multiaddrs).to.not.be.null;
 
-      const doesPeerIdExistInResponse =
-        peerInfos.find(
-          ({ ENR }) => ENR?.peerInfo?.id.toString() === nwaku1PeerId.toString()
-        ) !== undefined;
+      let foundNodeMultiaddrs: Multiaddr[] = [];
+      let foundNodePeerId: PeerId | undefined = undefined;
+      const doesPeerIdExistInResponse = peerInfos.some(({ ENR }) => {
+        foundNodeMultiaddrs = ENR?.peerInfo?.multiaddrs ?? [];
+        foundNodePeerId = ENR?.peerInfo?.id;
+        return ENR?.peerInfo?.id.toString() === nwaku1PeerId.toString();
+      });
+
+      if (!foundNodePeerId) {
+        throw new Error("Peer ID not found");
+      }
+
+      if (!foundNodePeerId) {
+        throw new Error("Peer ID not found");
+      }
 
       expect(doesPeerIdExistInResponse).to.be.equal(true);
 
-      expect(await waku.libp2p.peerStore.has(await nwaku2.getPeerId())).to.be
-        .true;
+      await waku.libp2p.dialProtocol(foundNodeMultiaddrs, PeerExchangeCodec);
+      await waitForRemotePeerWithCodec(
+        waku,
+        PeerExchangeCodec,
+        foundNodePeerId
+      );
+
+      expect(await waku.libp2p.peerStore.has(nwaku1PeerId)).to.eq(true);
+      expect(waku.libp2p.getConnections()).has.length(2);
     });
   });
 
@@ -87,12 +114,12 @@ describe("Peer Exchange", () => {
     this.timeout(55_000);
 
     let waku: LightNode;
-    let nwaku1: NimGoNode;
-    let nwaku2: NimGoNode;
+    let nwaku1: ServiceNode;
+    let nwaku2: ServiceNode;
 
     beforeEach(async function () {
-      nwaku1 = new NimGoNode(makeLogFileName(this) + "1");
-      nwaku2 = new NimGoNode(makeLogFileName(this) + "2");
+      nwaku1 = new ServiceNode(makeLogFileName(this) + "1");
+      nwaku2 = new ServiceNode(makeLogFileName(this) + "2");
     });
 
     tests({