+++ packages/tests/tests/relay/publish.node.spec.ts
@@ -1,18 +1,18 @@
 import { createEncoder } from "@waku/core";
 import { IRateLimitProof, RelayNode, SendError } from "@waku/interfaces";
-import { createRelayNode } from "@waku/sdk";
+import { createRelayNode } from "@waku/sdk/relay";
 import { utf8ToBytes } from "@waku/utils/bytes";
 import { expect } from "chai";
 
 import {
   delay,
+  generateRandomUint8Array,
   MessageCollector,
   NOISE_KEY_1,
   NOISE_KEY_2,
   tearDownNodes,
   TEST_STRING
 } from "../../src/index.js";
-import { generateRandomUint8Array } from "../../src/random_array.js";
 
 import {
   log,